# GrpcServer/openapi.yaml
openapi: 3.0.1
info:
  title: User & Group Management API
  description: RESTful CRUD API for Users and Groups with many-to-many relationship
  version: v1
servers:
  - url: https://localhost:7001
    description: Development server
  - url: http://localhost:5001
    description: Development server (HTTP)
paths:
  /users:
    get:
      tags:
        - Users
      summary: List all users
      operationId: GetUsers
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserDto'
    post:
      tags:
        - Users
      summary: Create a new user
      operationId: CreateUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserDto'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDto'
        '400':
          description: Bad Request
  /users/{userId}:
    get:
      tags:
        - Users
      summary: Get a specific user by ID
      operationId: GetUser
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDto'
        '404':
          description: Not Found
    put:
      tags:
        - Users
      summary: Replace an existing user
      operationId: ReplaceUser
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
            format: int32
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserDto'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDto'
        '400':
          description: Bad Request
        '404':
          description: Not Found
    patch:
      tags:
        - Users
      summary: Partially update a user
      operationId: PatchUser
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
            format: int32
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchUserDto'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDto'
        '400':
          description: Bad Request
        '404':
          description: Not Found
    delete:
      tags:
        - Users
      summary: Delete a user
      operationId: DeleteUser
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '204':
          description: No Content
        '404':
          description: Not Found
  /groups:
    get:
      tags:
        - Groups
      summary: List all groups
      operationId: GetGroups
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/GroupDto'
    post:
      tags:
        - Groups
      summary: Create a new group
      operationId: CreateGroup
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateGroupDto'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GroupDto'
        '400':
          description: Bad Request
  /groups/{groupId}:
    get:
      tags:
        - Groups
      summary: Get a specific group by ID
      operationId: GetGroup
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GroupDto'
        '404':
          description: Not Found
    put:
      tags:
        - Groups
      summary: Replace an existing group
      operationId: ReplaceGroup
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: integer
            format: int32
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateGroupDto'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GroupDto'
        '400':
          description: Bad Request
        '404':
          description: Not Found
    patch:
      tags:
        - Groups
      summary: Partially update a group
      operationId: PatchGroup
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: integer
            format: int32
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchGroupDto'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GroupDto'
        '400':
          description: Bad Request
        '404':
          description: Not Found
    delete:
      tags:
        - Groups
      summary: Delete a group
      operationId: DeleteGroup
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '204':
          description: No Content
        '404':
          description: Not Found
  /users/{userId}/groups:
    get:
      tags:
        - Relations
      summary: Get all groups that a user belongs to
      operationId: GetUserGroups
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/GroupDto'
    post:
      tags:
        - Relations
      summary: Add a user to one or more groups
      operationId: AddUserToGroups
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
            format: int32
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddUserToGroupsDto'
      responses:
        '204':
          description: No Content
        '400':
          description: Bad Request
  /users/{userId}/groups/{groupId}:
    delete:
      tags:
        - Relations
      summary: Remove a user from a specific group
      operationId: RemoveUserFromGroup
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
            format: int32
        - name: groupId
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '204':
          description: No Content
  /groups/{groupId}/users:
    get:
      tags:
        - Relations
      summary: Get all users in a specific group
      operationId: GetGroupUsers
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserDto'
    post:
      tags:
        - Relations
      summary: Add one or more users to a group
      operationId: AddUsersToGroup
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: integer
            format: int32
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddUsersToGroupDto'
      responses:
        '204':
          description: No Content
        '400':
          description: Bad Request
  /groups/{groupId}/users/{userId}:
    delete:
      tags:
        - Relations
      summary: Remove a specific user from a group
      operationId: RemoveUserFromGroupInGroupContext
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: integer
            format: int32
        - name: userId
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '204':
          description: No Content
components:
  schemas:
    UserDto:
      type: object
      required:
        - id
        - userName
        - email
      properties:
        id:
          type: integer
          format: int32
        userName:
          type: string
        email:
          type: string
          format: email
    CreateUserDto:
      type: object
      required:
        - userName
        - email
      properties:
        userName:
          type: string
          minLength: 3
        email:
          type: string
          format: email
    UpdateUserDto:
      type: object
      required:
        - userName
        - email
      properties:
        userName:
          type: string
          minLength: 3
        email:
          type: string
          format: email
    PatchUserDto:
      type: object
      properties:
        userName:
          type: string
          minLength: 3
          nullable: true
        email:
          type: string
          format: email
          nullable: true
    GroupDto:
      type: object
      required:
        - id
        - displayName
      properties:
        id:
          type: integer
          format: int32
        displayName:
          type: string
    CreateGroupDto:
      type: object
      required:
        - displayName
      properties:
        displayName:
          type: string
          minLength: 3
    UpdateGroupDto:
      type: object
      required:
        - displayName
      properties:
        displayName:
          type: string
          minLength: 3
    PatchGroupDto:
      type: object
      properties:
        displayName:
          type: string
          minLength: 3
          nullable: true
    AddUserToGroupsDto:
      type: object
      required:
        - groupIds
      properties:
        groupIds:
          type: array
          items:
            type: integer
            format: int32
    AddUsersToGroupDto:
      type: object
      required:
        - userIds
      properties:
        userIds:
          type: array
          items:
            type: integer
            format: int32